feat(langchaingo): add tests using testcontainers-go (#97)
This PR refactors the chat code to be more testable, using dependency injection of the OpenAI values (baseURL, key and model) in the form of function arguments. Therefore, it's possible to call the chat with differnt values (for production, or at test-time).

Once the code is more testable, this PR adds four different tests for the chat response:

using string comparison: it checks if the response contains certain string. Simple, although not very reliable.
using cosine similary and embeddings: it calculates the embeddings (numerical vectors) for the answer and compares them using the cosine similarity with the embeddings of a reference question.
using a vector database to store the "knowledge" or reference, doing RAG to augment the prompt with the relevant docs obtained from the vector database.
using an LLM-as-a-Judge: it creates an Evaluator, which is another model with a strict prompt, that evaluates the quality of the response of the chat LLM.
With these four tests, users can understand the different approaches to testing when building GenAI apps.